+++ chain/chain/src/spice/tests/mod.rs
@@ -1,3 +1,3 @@
-mod chunk_endorsements;
 mod core;
 mod core_writer_actor;
+mod header_chunk_endorsements;
